requirements-monitoring.txt
# Monitoring Requirements

# Metrics Collection
prometheus-client>=0.14.0
prometheus-flask-exporter>=0.20.0

# Tracing
opentelemetry-api>=1.14.0
opentelemetry-sdk>=1.14.0
opentelemetry-exporter-jaeger>=1.14.0
opentelemetry-instrumentation>=0.43b0
opentelemetry-instrumentation-flask>=0.43b0
opentelemetry-instrumentation-requests>=0.43b0
opentelemetry-instrumentation-sqlalchemy>=0.43b0

# Logging
structlog>=23.1.0
python-json-logger>=2.0.0

# Elasticsearch
elasticsearch>=8.0.0,<9.0.0
elasticsearch-py>=8.0.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0

# HTTP
requests>=2.28.0

# Configuration
python-dotenv>=0.20.0
pyyaml>=6.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
